app/requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.0.3
Flask-WTF==0.14.3
future==0.18.2
numpy==1.16.3
opencv-python==4.2.0.32
Pillow==6.1.0
requests==2.21.0
torch==1.5.1
torchvision==0.6.1